fixed tests paths
